move css files out of componenet folder
