+++ client/src/modal.css
