+++ client/src/photoGallery.css
